# public/openapi.yaml
openapi: 3.0.3
info:
  title: Norsk Tutor Public API
  version: "0.1.0"
  description: |
    Hand-maintained OpenAPI surface aligned with `lib/contracts/*` and `app/api/*` routes.
    Request correlation header `x-request-id` is accepted on all routes; session sync uses `x-device-id`.
servers:
  - url: /
paths:
...
  /api/openai-realtime:
    post:
      summary: Mint OpenAI Realtime session
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                model: { type: string }
                modalities:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: client_secret + session metadata
...